Move headers to a separate dir
