+++ product-service/handlers/helpers/commonHeaders.js
